# docs/registers/openapi.yaml
...
openapi: 3.0.3
info:
  title: RFS Station Manager API
  description: |
    REST API for the RFS Station Manager - a real-time digital sign-in system for Rural Fire Service stations.
    
    This API provides endpoints for:
    - Member management
    - Activity tracking
    - Check-in/check-out operations
    - Event management
    - Truck checks
    - Achievements
    
    Real-time updates are also available via Socket.io WebSocket connections.
  version: 1.0.0
  contact:
    name: RFS Station Manager Support
    url: https://github.com/richardthorek/Station-Manager
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://bungrfsstation.azurewebsites.net
    description: Production server
...
paths:
...
  /api/members:
    get:
      tags:
        - Members
      summary: Get all members
      description: Retrieve all registered members, sorted by activity level (check-ins in last 6 months) then alphabetically
      operationId: getMembers
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Members
      summary: Create new member
      description: Register a new member with auto-generated UUID and QR code
      operationId: createMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  example: John Smith
      responses:
        '201':
          description: Member created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/members/{id}:
    get:
      tags:
        - Members
      summary: Get member by ID
      description: Retrieve a specific member by their UUID
      operationId: getMemberById
      parameters:
        - name: id
          in: path
          required: true
          description: Member UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Members
      summary: Update member
      description: Update member information
      operationId: updateMember
      parameters:
        - name: id
          in: path
          required: true
          description: Member UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                rank:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Member updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/members/qr/{qrCode}:
    get:
      tags:
        - Members
      summary: Get member by QR code
      description: Retrieve a member by their unique QR code
      operationId: getMemberByQRCode
      parameters:
        - name: qrCode
          in: path
          required: true
          description: Unique QR code string
          schema:
            type: string
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/members/{id}/history:
    get:
      tags:
        - Members
      summary: Get member check-in history
      description: Retrieve all historical check-ins for a member
      operationId: getMemberHistory
      parameters:
        - name: id
          in: path
          required: true
          description: Member UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member check-in history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckIn'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/achievements/member/{memberId}:
    get:
      tags:
        - Achievements
      summary: Get member achievements
      description: Retrieve all achievements for a specific member
      operationId: getMemberAchievements
      parameters:
        - name: memberId
          in: path
          required: true
          description: Member UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Achievement'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/organizations/current/members:
    get:
      tags:
        - Organizations
      summary: List org members (multi-org memberships)
      operationId: listOrgMembers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Member list
...
components:
  schemas:
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        qrCode:
          type: string
        rank:
          type: string
          nullable: true
        memberNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - qrCode
        - createdAt
        - updatedAt
...
    CheckIn:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        activityId:
          type: string
          format: uuid
        checkInTime:
          type: string
          format: date-time
        checkOutTime:
          type: string
          format: date-time
          nullable: true
        checkInMethod:
          type: string
          enum: [kiosk, mobile, qr]
        location:
          type: string
          nullable: true
        isOffsite:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - activityId
        - checkInTime
        - checkInMethod
        - isOffsite
        - isActive
        - createdAt
        - updatedAt
...
    Achievement:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [sign_in, event, truck_check]
        requirement:
          type: integer
        icon:
          type: string
        isUnlocked:
          type: boolean
        unlockedAt:
          type: string
          format: date-time
          nullable: true
        progress:
          type: integer
      required:
        - id
        - name
        - description
        - category
        - requirement
        - icon
        - isUnlocked
        - progress
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    # Future: JWT Bearer token authentication for admin functions
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# No security required currently (by design for kiosk access)
# Future versions may require authentication for admin operations
security: []